Show new proposal button only if user has enough VP
